% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by.R
\name{aggregate_by}
\alias{aggregate_by}
\title{Select a field to aggregate results by}
\usage{
aggregate_by(request, field, limit = 1000)
}
\arguments{
\item{request}{string: request to modify}

\item{field}{string: name of field to aggregate results by}

\item{limit}{number: number of results to return. 1000 by default}
}
\value{
modified request as a list
}
\description{
Select a field to aggregate results by
}
