% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_data.R
\name{request_data}
\alias{request_data}
\title{Send a request for data}
\usage{
request_data(body)
}
\arguments{
\item{body}{string: list containing request}

\item{user}{string: username to authenticate request}

\item{pwd}{string: password to authenticate request}
}
\description{
Send a request for data
}
