% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_template.R
\name{request_template}
\alias{request_template}
\title{Return a nested list containing the base request template}
\usage{
request_template(path)
}
\arguments{
\item{path}{string: path specifiying where the json template is stored}
}
\value{
list containing base request information
}
\description{
Return a nested list containing the base request template
}
