% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squash_buckets.R
\name{squash_buckets}
\alias{squash_buckets}
\title{Turn nested json into 2D dataframe}
\usage{
squash_buckets(result)
}
\arguments{
\item{resp}{string: response object to convert}
}
\value{
data.frame of data
}
\description{
Handles case of one or two aggregations
}
